add scripts to collect deta and fix 3 segmentation fault bug in file 7,8,11 cuda

+++ main_naif_cpu_2.cpp
@@ -130,9 +130,9 @@ void calculateParameters(const std::vector<float> &prices, float &S0, float &dri
 // --- MAIN ---
 int main(int argc, char *argv[])
 {
-    auto elapsedSort = std::chrono::duration<float, std::milli>::zero();
-    auto elapsedMC = std::chrono::duration<float, std::milli>::zero();
-    auto elapsedParam = std::chrono::duration<float, std::milli>::zero();
+    auto elapsedSort = std::chrono::duration<double, std::milli>::zero();
+    auto elapsedMC = std::chrono::duration<double, std::milli>::zero();
+    auto elapsedParam = std::chrono::duration<double, std::milli>::zero();
 
 // Valore di default se l'utente non inserisce argomenti
     long nSimulations = 10000000;
